Introduce new webhooks API
