+++ app/config/scopes/project.php
@@ -172,4 +172,12 @@ return [ // List of publicly visible scopes
     'tokens.write' => [
         'description' => 'Access to create, update, and delete your project\'s tokens',
     ],
+    "webhooks.read" => [
+        "description" =>
+            "Access to read project\'s webhooks",
+    ],
+    "webhooks.write" => [
+        "description" =>
+            "Access to create, update, and delete project\'s webhooks",
+    ],
 ];
