+++ src/Appwrite/Platform/Appwrite.php
@@ -16,6 +16,7 @@ use Appwrite\Platform\Modules\Storage;
 use Appwrite\Platform\Modules\Teams;
 use Appwrite\Platform\Modules\Tokens;
 use Appwrite\Platform\Modules\VCS;
+use Appwrite\Platform\Modules\Webhooks;
 use Utopia\Platform\Platform;
 
 class Appwrite extends Platform
@@ -36,5 +37,6 @@ class Appwrite extends Platform
         $this->addModule(new Tokens\Module());
         $this->addModule(new Storage\Module());
         $this->addModule(new VCS\Module());
+        $this->addModule(new Webhooks\Module());
     }
 }
