+++ src/Appwrite/Platform/Modules/Webhooks/Module.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Appwrite\Platform\Modules\Webhooks;
+
+use Appwrite\Platform\Modules\Webhooks\Services\Http;
+use Utopia\Platform;
+
+class Module extends Platform\Module
+{
+    public function __construct()
+    {
+        $this->addService('http', new Http());
+    }
+}
