+++ src/Appwrite/Platform/Workers/Migrations.php
@@ -332,6 +332,8 @@ class Migrations extends Action
                 'messages.write',
                 'targets.read',
                 'targets.write',
+                'webhooks.read',
+                'webhooks.write'
             ]
         ]);
