+++ tests/e2e/Services/Webhooks/WebhooksConsoleClientTest.php
@@ -4,11 +4,11 @@ namespace Tests\E2E\Services\Webhooks;
 
 use Tests\E2E\Scopes\ProjectCustom;
 use Tests\E2E\Scopes\Scope;
-use Tests\E2E\Scopes\SideServer;
+use Tests\E2E\Scopes\SideConsole;
 
-class WebhooksCustomServerTest extends Scope
+class WebhooksConsoleClientTest extends Scope
 {
     use WebhooksBase;
     use ProjectCustom;
-    use SideServer;
+    use SideConsole;
 }
